Fix: 코드 수정 및 연결 기능 추가
