+++ src/main/kotlin/com/noreabang/strawberryrabbit/domain/comment/dto/CommentRequest.kt
@@ -1,5 +1,5 @@
 package com.noreabang.strawberryrabbit.domain.comment.dto
 
 data class CommentRequest(
-    val content: String,
+    val content: String?,
 )
