+++ src/main/kotlin/com/noreabang/strawberryrabbit/domain/comment/repository/CommentRepository.kt
@@ -4,5 +4,4 @@ import com.noreabang.strawberryrabbit.domain.comment.model.Comment
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface CommentRepository: JpaRepository<Comment, Long> {
-
 }
